Add decorator for Mako filters to disable defaults for the context
To this point, the Mako filter ``sjson`` required that the user remember
to also apply the ``n`` filter with each usage. Example:

``var jsVar = ${ python_var | sjson, n };``

With the addition of this decorator, apply the sjson filter will turn
off the default filters for the context automatically. The above code
may now be written as:

``var jsVar = ${ python_var | sjson };``

Note that the additional application of the ``n`` filter now has no
effect - existing code that applies both filters need not be changed.